manifest: condense a glob into a dependency
Pass a list of files on stdin, and the manifest tool will stat each of
them, updating a manifest file with the contents of which files changed
and ensuring that the manifest file's mod time is always at least as
recent as the most recently modified file in the list.

The manifest file is then usable as an input to a build rule.