switch to _Module-based solution
The _Module object will collect groups of targets based on the
module in which they are defined.  This will allow for all-like targets
at the _Project level as well as within any _Module level.

This also eliminates all the global state, except for _aliases which
applies to the import system (which is already built on singletons).

As a side-effect, add_target_object is now an (undocumented) symbol
available in all mkninja.py files, because its implementation is tied to
a _Module.