support add_series()
The project-wide ALL series is the default series for any target.

Additional series may be created with add_series().

I don't know if I like the extra complication added by this.  When a
particular directory has just one target, like say a "lint" target, it
would be much more natural I think to just have a regular target that
names BLD/"lint" as its output, than to have a target with no name but
series=root.LINT which magically gets pulled into the `subdir/lint`
target.

I think this feature will be left undocumented for a while, and maybe
removed in the future.