default=False, stamp=True, and add_alias()
default=False replaces the add_series() feature.  add_series() was too
complicated to remember how it worked, and it introduced weird
complexities, like "why can't the test series depend on the build
series"?  In practice, the main thing I really wanted it for was to make
some things not default.  Everything else is effectively covered by
using explicit targets (like a "lint" target in every directory, for
instance).

stamp=True is a new primitive that causes the first output for the
target to be treated as a stamp file automatically.  This saves a few
lines of code every time a stamp file is needed.

add_alias() is another new primitive, which makes many stamp files
unnecessary.